migrate ExitCondition to separate file
